Project/requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
importlib-metadata
Markdown
Pillow
requests
requests-oauthlib
asgiref
coverage
flake8
mccabe
pycodestyle
pyflakes
pytz
sqlparse
channels
django-cors-headers